refactor: move update command to modle
